EffectQueue: promote Globals from file scope, so the mapping can be reset
fixes not being able to run more than one game test: the demo loaded its
dummy data for MapTest and then no effects resolved for subsequent games

the same is true with more than one resave test, but that still crashes
elsewhere for some combinations of tests. Oddly individual tests can give
different results if run together
